Send camera 0 through protobuf means no change
As protobuf 3 doesn't allow to set default camera value to -1 (it's not the
default for its type, so 0), we have to offset by 1 for the client all
camera numbers.
It's also better to read "Camera 1, Camera 2" than "Camera 0, Camera 1" even
if this it the internal opencv representation, so it's a win.